backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crewai>=0.175.0",
    "fastapi>=0.116.1",
    "openai>=1.102.0",
    "python-dotenv>=1.1.1",
    "supabase>=2.18.1",
    "uvicorn>=0.35.0",
]
